Create AutoSplineCommand
Command that uses vision to select a node, generate a spline to that node, and then follow the spline.

+++ Robot2023/src/main/java/com/team2813/lib/util/LimelightValues.java
@@ -82,8 +82,8 @@ public class LimelightValues {
     }
 
    /** Id of the primary Apriltag */
-    public double primaryApriltag() {
-        return primaryApriltag.get();
+    public int primaryApriltag() {
+        return primaryApriltag.get().intValue();
     } 
 
     public double getTx() {
